Get codebase completely working with LLVM
You can now build Cosmopolitan with Clang:

    make -j8 MODE=llvm
    o/llvm/examples/hello.com

The assembler and linker code is now friendly to LLVM too.
So it's not needed to configure Clang to use binutils under
the hood. If you love LLVM then you can now use pure LLVM.

+++ libc/rand/randf.c
@@ -22,6 +22,6 @@
 #include "libc/rand/rand.h"
 
 float randf(void) {
-  return (float)(int)(KnuthLinearCongruentialGenerator(&g_rando) >> 32) /
+  return (double)(int)(KnuthLinearCongruentialGenerator(&g_rando) >> 32) /
          INT_MAX;
 }
